EnergyCampus: fix computation of consumption per  squaremeter
Ennovatis seems to provide wrong data via REST interface for channels per squaremeter.

Workaround: Use historical dataset to get the floor space from GIS and compute it (total consumption / floor space)

+++ src/OpenCOVER/plugins/hlrs/Energy/ennovatis/json.h
@@ -14,7 +14,7 @@ struct json_response_object {
   int MinValue=0;
   int StandardDeviation=0;
   std::vector<std::string> Times;
-  std::vector<int> Values;
+  std::vector<float> Values;
   operator std::string() const;
 };
